Admin Panel/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyPDF2>=3.0.1
APScheduler>=3.10.4
python-dateutil>=2.8.2
aiosqlite>=0.20.0
pdfminer.six>=20221105
bleach>=6.2.0
pytest>=8.3.0
httpx>=0.28.0
